Add a diskcache for Mapnik(=openstreetmap)
The loading of tiles from openstreetmap is sometimes slow.
Add a new provider MapnikDiskCache, which caches tiles on disk.
Tiles are stored under ~/HOME/.cache/openstreetmap

TODO: The tiles never expire, so that the user needs to remove all
the cache when a refresh is needed.

Beside that, the following changes/cleanups are included:
OsmTileLoader.java: Rename loadTileFromOsm() into getUrlConnection()
    The method loadTileFromOsm() does not load anything, but returns
    an UrlConnection which is used for the download.
    Rename it.

Use baseUrl when appropriate
    In some places the term url is used, but the variable named url only
    store the baseUrl. Chane them into baseUrl when needed to make clear what is going on