Add UI component and filter to set visibility of multiple GPXObjects
Before this feature, the only way to set the visibility of a GPXObject was by clicking on the tiny lamp icon on each GPXObject in the "Explorer" panel. With this new feature, it is possible to show or hide multiple GPXObject at once. By entering a keyword in the search bar, all entries which contain the search query will be hidden or shown.