Fix GpsMaster/createJAR.sh
We still have an old way to build with an old JDK 8.
The binary will run under JDK8 (possibly 7), 9, 10.
Everybody using Java 11 and higher should use the binary for JDK11,
which is generated with mvn.
Before we kick out the old JDK8 for good, fix the script and
postpone the cleanup a little bit.