GpsMaster.java: Improve lat- and lon input fields
You can place the "cursor" on the map by entering latitude and longitude
into the 2 fields on top of the mapping.
The coordinates must be in double format.

Improve the situation, allow stings like "51° 28′ 38″ N" and "0" to place
the cursor in Greenwich.

The current parser supports 2 format types:
"degree minutes seconds" or the simple double format.
seconds may have decimals (or not).
The separaotor between degree minutes and minutes seconds is relaxed,
anything with is not a digit is accepted.
Like `51°28'38"` or `51° 28' 38"` or other versions.