Hike and Bike Map, OpenTopMap: Refactor code out of GpsMaster.java
Factor out the main code for "Hike & Ride Map" and OpenTopMap
from GpsMaster.java into
src/org/openstreetmap/gui/jmapviewer/tilesources/OsmTileSource.java

This makes the code a little bit more consistent:
Even if they are not OSM as such, they follow the same principle.
(And after this change, it is easier to add a disk-cache for them.