OsmTileLoader.java: Refresh the disk cache
There is a kind of dilemma:
The current implementation is never updating the disk-cache.

This is good and bad:
- If you have a track from 10 years ago, and re-read it,
  there is a good chance that you get the map from the cache.

- If you have a track from "today", this data may be out-dated,
  and "OSM Online" must be used.

Take a first step to improve the situation:
Refresh the disk cache when the tile has the age of 4 weeks.
This gives still a good balance between "old" and "new" data.

Note: When we are offline, the tile in the disk cache may be old,
but since we are offline, this is all that we have.
Use the "old" tile instead of showing nothing.

TODO:
Introduce a kind of "historic" cache. Lets say, we use a new cache
once a year. That will save the historic data in an own folder,
and opens the possiblity to be re-used whenever we load an old track.
More implementations are needed, so lets commit this "as is".