OsmTileSource.java: Disk cache invisible in user menu
The Open Street Map (OSM) is selectable in 2 versions:
- "OSM Online"
- "OSM DiskCache"

The Online version does what is promises: It fetches tile online and
stores them in memory as long as GpsMaster is running.
The DiskCache version reads the tile from cache, if it is in the cache.
If it is not in the cache, the tile is retrieved online and
added into the cache. This cache never expires.
If there is a route, where we have values in the cache, they stay there
forever. When the route uses a road that does not exist any more,
you still see it from the cache.
Swithing to "OSM Online" may show a different image:
You may have driven through a house, a forest...

I don't know how to solve the dilemma that we need old historic data
and new fresh data.
For the moment, we keep it as it is for OSM.

Then there are 2 new providers:
- CycleMapDiskCache :       "CycleMap"
- HikeAndBikeMapDiskCache : "Hike&Bike"

Those 2 use a disk cache as well, which is refreshed after 28 days.