dialogs/LatLonPanel.java: Allow copy-paste of combined coordinates
Allow copy-and-paste into the lat field.
Copying coordinates from www.openstreetmap.org is now easier:
The URL ends with something like map=14/52.5100/13.3921

Allow to paste the coordinates "52.5100/13.3921" into the lat field:
Try to find a '/' in the string, and then split it into lat and lon.

Add another regexp:
"52.5100 13.3921" with only one space.