LatLonPanel.java: Allow lat,lon trim() the strings
When copy-pasting stings from internet, it may happen that leading
or trailing spaces are copied as well.
However, they make using a regexp that use of ' ' as a seperator
harder than needed.
Solution: trim() the string(s)

Add another "combined format" of lat and lon:
12.345, 56.789
The seperator  is comma and/or space, which means that the coordinates
must use a decimal-dot